LayoutTests/imported/w3c/csswg-test/css-shapes-1/shape-outside/values/shape-outside-polygon-003-expected.txt


FAIL polygon() - invalid assert_equals: expected (object) null but got (string) ""
FAIL polygon(evenodd) - invalid assert_equals: expected (object) null but got (string) ""
FAIL polygon(nonzero) - invalid assert_equals: expected (object) null but got (string) ""
FAIL polygon(evenodd, foo bar) - invalid assert_equals: expected (object) null but got (string) ""
FAIL polygon(nonzero, bar foo) - invalid assert_equals: expected (object) null but got (string) ""
FAIL polygon(foo, 1px 2px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL polygon(evenodd 1px 2px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL polygon(nonzero 1px 2px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL polygon(1px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL polygon(1px 2px 3px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL polygon(1px 2px 3px 4px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL polygon(1px, 2px) - invalid assert_equals: expected (object) null but got (string) ""
FAIL polygon(1px 2px evenodd) - invalid assert_equals: expected (object) null but got (string) ""
FAIL polygon(1px 2px nonzero) - invalid assert_equals: expected (object) null but got (string) ""
FAIL polygon(evenodd, 10 20) - invalid assert_equals: expected (object) null but got (string) ""
FAIL polygon(nonzero, 10 20) - invalid assert_equals: expected (object) null but got (string) ""
